main: improve cleanup mechanism
This commit improves the cleanup mechanism that runs in the event of
errors. It now includes not only PKCS11 keys and certificates but also
files stored locally, and the registration from a back-end point of
view. The cleanup process is now executed in any error situation during
execution, and also when SIGINT and SIGSEGV signals are received.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>